REF: Create jackknife namespace
